#1 - refactor teste data from user to its own folder
